Fix colorspace display bug
The implementation of the display should not be a defaultdict because
that will add new values to the map when they are encountered, which is
undesired behavior.  Change the implementation to a regular dict.

The test fixtures for TestPrinting seem wrong.  The options passed to
set_printoptions seem like they should be the same for setup as for
teardown.